requirements.txt
pyTelegramBotAPI>=4.32.0
python-dotenv>=1.0.0
